refactor(Parse): split Parse into multiple files
